Remove dead drift detection, rename drift.ts to utils.ts
