--- clients/admin-ui/src/features/properties/privacy-center-config/form-builder/__tests__/detectSpecPcShapeDrift.test.ts
@@ -1,49 +1,0 @@
-import { detectSpecPcShapeDrift } from "../detectSpecPcShapeDrift";
-
-describe("detectSpecPcShapeDrift", () => {
-  it("returns false when the rich spec maps to the saved PC shape", () => {
-    const richSpec = {
-      root: "form",
-      elements: {
-        form: { type: "Form", props: {}, children: ["f_email"] },
-        f_email: {
-          type: "Text",
-          props: { name: "email", label: "Email", required: true },
-          children: [],
-        },
-      },
-    };
-    const savedPc = {
-      email: {
-        label: "Email",
-        field_type: "text" as const,
-        required: true,
-      },
-    };
-
-    expect(detectSpecPcShapeDrift(richSpec, savedPc)).toBe(false);
-  });
-
-  it("returns true when fields differ", () => {
-    const richSpec = {
-      root: "form",
-      elements: {
-        form: { type: "Form", props: {}, children: ["f_email"] },
-        f_email: {
-          type: "Text",
-          props: { name: "email", label: "Email", required: true },
-          children: [],
-        },
-      },
-    };
-    const savedPc = {
-      email: {
-        label: "Email",
-        field_type: "text" as const,
-        required: false,
-      },
-    };
-
-    expect(detectSpecPcShapeDrift(richSpec, savedPc)).toBe(true);
-  });
-});
